+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The **patch** part changes incrementally at each release.
 * listupgrade: crontab is configurable
 * mongodb: create munin plugins directory if missing
 * mysql: script "mysql_connections" to display a compact list of connections
+* mysql: script "mysql-queries-killer.sh" to kill MySQL queries
 * nagios-nrpe + evolinux-users: new checks for bkctld
 * redis: instance service for Debian 11